Refactoring source files to use snake_case
