+++ lib/rotor_utils.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_blue/flutter_blue.dart';
-import 'package:mobileclient/Strings.dart';
+import 'package:mobileclient/strings.dart';
 
 class RotorUtils {
