+++ lib/strings.dart
