+++ lib/ui/commonwidgets/device_row.dart
