+++ lib/ui/vehiclemonitor/vehicle_info_page_content.dart
