+++ lib/ui/vehiclemonitor/vehicle_metrics_page_content.dart
