+++ lib/ui/welcome/welcome_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mobileclient/ui/welcome/WelcomePageContent.dart';
+import 'package:mobileclient/ui/welcome/welcome_page_content.dart';
 
 class WelcomePage extends StatelessWidget {
   @override
