+++ test/ui/commonwidgets/bt_connection_dialog_widget_test.dart
@@ -4,7 +4,7 @@ import 'package:flutter_blue/flutter_blue.dart';
 import 'package:mockito/mockito.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mobileclient/ui/commonwidgets/BTConnectionDialog.dart';
+import 'package:mobileclient/ui/commonwidgets/bt_connection_dialog.dart';
 
 import '../../mocks/rotor_mocks.dart';
