+++ test/ui/devicelist/device_list_page_content_state_test.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_blue/flutter_blue.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mobileclient/Strings.dart';
+import 'package:mobileclient/strings.dart';
 import 'package:mockito/mockito.dart';
-import 'package:mobileclient/ui/devicelist/DeviceListPageContent.dart';
+import 'package:mobileclient/ui/devicelist/device_list_page_content.dart';
 
 import '../../mocks/rotor_mocks.dart';
