+++ test/ui/devicelist/device_list_page_content_widget_test.dart
@@ -3,9 +3,9 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_blue/flutter_blue.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mobileclient/Strings.dart';
-import 'package:mobileclient/ui/commonwidgets/Notice.dart';
-import 'package:mobileclient/ui/devicelist/DeviceListPageContent.dart';
+import 'package:mobileclient/strings.dart';
+import 'package:mobileclient/ui/commonwidgets/notice.dart';
+import 'package:mobileclient/ui/devicelist/device_list_page_content.dart';
 import 'package:mockito/mockito.dart';
 
 import '../../mocks/rotor_mocks.dart';
