+++ test/ui/vehiclemonitor/vehicle_controls_page_content_widget_test.dart
@@ -3,8 +3,8 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_blue/flutter_blue.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mobileclient/data/RotorCommand.dart';
-import 'package:mobileclient/ui/vehiclemonitor/VehicleControlsPageContent.dart';
+import 'package:mobileclient/data/rotor_command.dart';
+import 'package:mobileclient/ui/vehiclemonitor/vehicle_controls_page_content.dart';
 import 'package:mockito/mockito.dart';
 
 import '../../mocks/rotor_mocks.dart';
